requirements.txt
adal==1.2.2
alembic==1.0.11
aniso8601==7.0.0
asn1crypto==0.24.0
astroid==2.2.5
attrs==19.1.0
azure-applicationinsights==0.1.0
azure-batch==4.1.3
azure-common==1.1.23
azure-cosmosdb-nspkg==2.0.2     
azure-cosmosdb-table==1.0.5
azure-datalake-store==0.0.46
azure-graphrbac==0.40.0
azure-keyvault==1.1.0
azure-loganalytics==0.1.0
azure-mgmt-advisor==1.0.1
azure-mgmt-applicationinsights==0.1.1
azure-mgmt-batchai==2.0.0
azure-mgmt-cdn==3.1.0
azure-mgmt-commerce==1.0.1
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==1.5.0
azure-mgmt-datalake-analytics==0.6.0
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datamigration==1.0.0
azure-mgmt-devtestlabs==2.2.0
azure-mgmt-eventhub==2.6.0
azure-mgmt-hanaonazure==0.1.1
azure-mgmt-iotcentral==0.1.0
azure-mgmt-iothubprovisioningservices==0.2.0
azure-mgmt-machinelearningcompute==0.4.1
azure-mgmt-managementgroups==0.1.0
azure-mgmt-managementpartner==0.1.1
azure-mgmt-maps==0.1.0
azure-mgmt-marketplaceordering==0.1.0
azure-mgmt-media==1.0.0
azure-mgmt-msi==0.2.0
azure-mgmt-network==2.7.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==0.1.0
azure-mgmt-powerbiembedded==2.0.0
azure-mgmt-rdbms==1.9.0
azure-mgmt-recoveryservicesbackup==0.3.0
azure-mgmt-resource==2.2.0
azure-mgmt-scheduler==2.0.0
azure-mgmt-search==2.1.0
azure-mgmt-servicefabric==0.2.0
azure-mgmt-signalr==0.1.1
azure-mgmt-storage==2.0.0
azure-mgmt-subscription==0.2.0
azure-mgmt-trafficmanager==0.50.0
azure-nspkg==3.0.2
azure-servicebus==0.21.1
azure-servicefabric==6.3.0.0
azure-servicemanagement-legacy==0.20.6
azure-storage-blob==1.5.0
azure-storage-common==1.4.2
azure-storage-file==1.4.0
azure-storage-queue==1.4.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.7
decorator==4.4.0
ffmpeg==1.4
Flask==1.1.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
flask-restplus==0.12.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
idna==2.8
imageio==2.5.0
imageio-ffmpeg==0.3.0
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
lazy-object-proxy==1.4.1
Mako==1.0.14
MarkupSafe==1.1.1
mccabe==0.6.1
moviepy==1.0.0
msrest==0.6.8
msrestazure==0.6.1
numpy==1.16.4
oauthlib==3.0.2
opencv-python==3.4.5.20
Pillow==6.1.0
proglog==0.1.9
pycparser==2.19
pygame==1.9.6
PyJWT==1.7.1
pylint==2.3.1
pyrsistent==0.15.4
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
SQLAlchemy==1.3.6
tqdm==4.32.2
typed-ast==1.4.0
urllib3==1.25.3
Werkzeug==0.15.5
wrapt==1.11.2
WTForms==2.2.1
youtube-dl==2019.6.27